Fix affine warping with openCV for number of classes > 4
When using imgaug to perform paired augmentation of an image and of its  segmentation map, the segmentation map is converted to 1-hot encoding. This can create images with more than 4 channels, depending on the number of classes. OpenCV cannot always directly perform the warping on such images, it should be applied for each channel independently.